lots of bug fixes and improvements  - used chrono::duration in sleep_for function, better readability  - changed data type no mixing (remove unsigned and replaced float with double)  - renamed a couples of class to better represent its job and added some comments for class methods  - used a new data structure for time, so less specical case is needed for schedule checking.  - added code to output detected contour area for parameter tuning

+++ src/libs/DateTime/src/Timer.cpp
@@ -13,10 +13,10 @@ void Timer::initialize() {
 void Timer::reset() {
   t0 = clock();
 }
-float Timer::get_dt_msec()
+double Timer::get_dt_msec()
 {
   return get_dt_msec()*1000.0;
 }
-float Timer::get_dt_sec() {
-  return ((float)clock() - (float)t0) / CLOCKS_PER_SEC;
+double Timer::get_dt_sec() {
+  return ((double)clock() - (double)t0) / (double) CLOCKS_PER_SEC;
 }
